chore: move `parse_str_to_field` into `mod.rs`
